Rename EndpointTestTrait and adjust testing of it (#12)
